bgp-contd: too early to check connectivity to r20
The first thing I did in this assignment is disabled connection
R1-R20. And spend some time trying to understand how it is even
supposed to work, before I scrolled down and read that it is not
supposed actually.

So let's not ask a person to check if he can reach ALL parts of the
network, but only reasonable for now parts.